requirements.txt
Flask>=2.2.0
python-telegram-bot==20.7
httpx>=0.26.0
